+++ src/modules/auth/auth.controller.ts
@@ -1,9 +1,7 @@
-import { Controller, Post, Body, Get, HttpCode, HttpStatus, Res, Query } from '@nestjs/common';
+import { Controller, Post, Body, HttpCode, HttpStatus, Res } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthLoginDto } from './dto/auth.login.dto';
 import { CreateUserDto } from './dto/auth.register.dto';
-import { ForgotPasswordDto } from './dto/forgot-password.dto';
-import { ResetPasswordDto } from './dto/reset-password.dto';
 
 @Controller('auth')
 export class AuthController {
@@ -18,16 +16,6 @@ export class AuthController {
   @HttpCode(HttpStatus.OK)
   async login(@Body() loginDto: AuthLoginDto) {
     return await this.authService.signIn(loginDto);
-  }
-
-  @Post('forgot-password')
-  async forgotPassword(@Body() forgotPasswordDto: ForgotPasswordDto) {
-    return await this.authService.requestPasswordReset(forgotPasswordDto);
-  }
-
-  @Post('reset-password')
-  async resetPassword(@Query('token') token: string, @Body() resetPasswordDto: ResetPasswordDto) {
-    return await this.authService.resetPassword(token, resetPasswordDto);
   }
 }
