+++ src/modules/auth/auth.service.ts
@@ -1,13 +1,8 @@
-import { ConflictException, Injectable, InternalServerErrorException, NotFoundException, UnauthorizedException,  BadRequestException } from '@nestjs/common';
+import { ConflictException, Injectable, InternalServerErrorException, UnauthorizedException} from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { AuthLoginDto } from './dto/auth.login.dto';
 import { PrismaService } from '../../prisma.service';
 import { CreateUserDto } from './dto/auth.register.dto';
-import { ForgotPasswordDto } from './dto/forgot-password.dto';
-import { MailService } from '../mail/mail.service';
-import { ResetPasswordDto } from './dto/reset-password.dto';
-import { v4 as uuidv4 } from 'uuid';
-import { User } from '@prisma/client';
 import * as bcrypt from 'bcrypt';
 
 @Injectable()
@@ -15,7 +10,6 @@ export class AuthService {
   constructor(
     private prisma: PrismaService,
     private jwtService: JwtService,
-    private mailService: MailService
   ) { }
 
 
@@ -105,62 +99,4 @@ export class AuthService {
       throw error
     }
   }
-
-  async requestPasswordReset(forgotPasswordDto: ForgotPasswordDto) {
-    const { email } = forgotPasswordDto;
-  
-    try {
-      const user = await this.prisma.user.findUnique({ where: { email } });
-      if (!user) {
-        throw new NotFoundException('Usuario no encontrado');
-      }
-  
-      const token = this.jwtService.sign({ userId: user.id }, { expiresIn: '2h' });
-      await this.mailService.sendResetPasswordEmail(user.email, token);
-  
-      return { message: 'Correo de recuperación enviado' };
-    } catch (error) {
-      throw new InternalServerErrorException('Ocurrió un error al procesar tu solicitud.');
-    }
-  }
-
-  async verifyToken(token: string): Promise<User> {
-    try {
-        const decoded = this.jwtService.verify(token);
-        const user = await this.prisma.user.findUnique({ where: { id: decoded.userId } });
-        if (!user) {
-            throw new NotFoundException('Usuario no encontrado');
-        }
-        return user;
-    } catch (error) {
-        throw new UnauthorizedException('Token inválido o expirado.');
-    }
-}
-
-  async resetPassword(token: string, resetPasswordDto: ResetPasswordDto) {
-  try {
-    const decodedUser = await this.verifyToken(token);
-    
-    const user = await this.prisma.user.findUnique({ where: { id: decodedUser.id } });
-
-    if (!user) {
-      throw new NotFoundException('Usuario no encontrado.');
-    }
-
-    if (resetPasswordDto.newPassword !== resetPasswordDto.repeatPassword) {
-      throw new BadRequestException('Las contraseñas no coinciden.');
-    }
-
-    const hashedPassword = await bcrypt.hash(resetPasswordDto.newPassword, 10);
-
-    await this.prisma.user.update({
-      where: { id: user.id },
-      data: { password: hashedPassword },
-    });
-
-    return { message: 'Contraseña actualizada correctamente.' };
-  } catch (error) {
-    throw new InternalServerErrorException('Ocurrió un error al procesar la solicitud de restablecimiento de contraseña.');
-  }
-}
 }
