+++ src/modules/mail/mail.service.ts
@@ -1,44 +1,118 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, InternalServerErrorException, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import { ForgotPasswordDto } from './dto/forgot-password.dto';
+import { ResetPasswordDto } from './dto/reset-password.dto';
+import { PrismaService } from '../../prisma.service';
+import { JwtService } from '@nestjs/jwt';
+import { User } from '@prisma/client';
+import * as bcrypt from 'bcrypt';
 import * as nodemailer from 'nodemailer';
 import * as fs from 'fs';
-import * as path from 'path'
-
+import * as path from 'path';
 
 @Injectable()
 export class MailService {
   private transporter: nodemailer.Transporter;
-  
 
-  constructor() {
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly jwtService: JwtService
+  ) {
     this.transporter = nodemailer.createTransport({
-      service: 'Gmail', 
+      service: 'Gmail',
       auth: {
-        user: process.env.EMAIL_USER, 
+        user: process.env.EMAIL_USER,
         pass: process.env.EMAIL_PASSWORD,
       },
     });
   }
 
   async sendResetPasswordEmail(to: string, token: string) {
-    const resetLink = `${process.env.FRONTEND_URL}/auth/reset-password?token=${token}`;
+    const resetLink = `${process.env.FRONTEND_URL}/restore-password/reset-password?token=${token}`;
     const templatePath = path.join(process.cwd(), 'src', 'utils', 'emailTemplate.html');
 
     let emailTemplate = fs.readFileSync(templatePath, 'utf-8');
     emailTemplate = emailTemplate.replace('{{resetLink}}', resetLink);
-    
+
     const mailOptions = {
-        from: 'EventMap',
-        to,
-        subject: 'Recuperación de contraseña',
-        html: emailTemplate,
+      from: 'EventMap',
+      to,
+      subject: 'Recuperación de contraseña',
+      html: emailTemplate,
     };
 
     try {
-        await this.transporter.sendMail(mailOptions);
+      await this.transporter.sendMail(mailOptions);
     } catch (error) {
-        console.error('Error al enviar correo electrónico:', error);
-        throw new Error('Error enviando el correo');
+      console.error('Error al enviar correo electrónico:', error);
+      throw new Error('Error enviando el correo');
     }
-}
+  }
 
+  async requestPasswordReset(forgotPasswordDto: ForgotPasswordDto) {
+    const { email } = forgotPasswordDto;
+    console.log('Correo electrónico recibido para el restablecimiento de contraseña:', email);
+
+    try {
+      const user = await this.prisma.user.findUnique({ where: { email } });
+
+      if (!user) {
+        console.log('Usuario no encontrado para el correo:', email);
+        throw new NotFoundException('Usuario no encontrado');
+      }
+
+      console.log('Usuario encontrado:', user);
+
+      const token = this.jwtService.sign({ userId: user.id }, { expiresIn: '2h' });
+      console.log('Token generado:', token);
+
+      // No necesitas usar `this.mailService`, ya estás en el mismo servicio
+      await this.sendResetPasswordEmail(user.email, token);
+      console.log('Correo de recuperación enviado a:', user.email);
+
+      return { message: 'Correo de recuperación enviado' };
+    } catch (error) {
+      console.error('Error en la solicitud de restablecimiento de contraseña:', error);
+      throw new InternalServerErrorException('Ocurrió un error al procesar tu solicitud.');
+    }
+  }
+
+  async verifyToken(token: string): Promise<User> {
+    try {
+      const decoded = this.jwtService.verify(token);
+      const user = await this.prisma.user.findUnique({ where: { id: decoded.userId } });
+      if (!user) {
+        throw new NotFoundException('Usuario no encontrado');
+      }
+      return user;
+    } catch (error) {
+      throw new UnauthorizedException('Token inválido o expirado.');
+    }
+  }
+
+  async resetPassword(token: string, resetPasswordDto: ResetPasswordDto) {
+    try {
+      const decodedUser = await this.verifyToken(token);
+
+      const user = await this.prisma.user.findUnique({ where: { id: decodedUser.id } });
+
+      if (!user) {
+        throw new NotFoundException('Usuario no encontrado.');
+      }
+
+      if (resetPasswordDto.newPassword !== resetPasswordDto.repeatPassword) {
+        throw new BadRequestException('Las contraseñas no coinciden.');
+      }
+
+      const hashedPassword = await bcrypt.hash(resetPasswordDto.newPassword, 10);
+
+      await this.prisma.user.update({
+        where: { id: user.id },
+        data: { password: hashedPassword },
+      });
+
+      return { message: 'Contraseña actualizada correctamente.' };
+    } catch (error) {
+      throw new InternalServerErrorException('Ocurrió un error al procesar la solicitud de restablecimiento de contraseña.');
+    }
+  }
 }
